--- aries-site/src/data/structures/extend.js
@@ -1,24 +1,0 @@
-/* eslint-disable max-len */
-export const extend = [
-  {
-    name: 'Roadmap',
-    available: true,
-    description:
-      'An interactive roadmap of the HPE Design System that provides insight into current and future milestones.',
-    preview: {
-      image: {
-        src: {
-          light: '/extendImages/extend-preview-roadmap.svg',
-          dark: '/extendImages/extend-preview-roadmap-invert.svg',
-        },
-        alt: 'HPE Design System Roadmap',
-        fit: 'contain',
-      },
-    },
-    seoDescription:
-      'An interactive roadmap of the HPE Design System that provides insight into current and future milestones.',
-    sections: [],
-    url:
-      'https://github.com/orgs/grommet/projects/11?fullscreen=true',
-  },
-];
