+++ aries-site/src/data/structures/index.js
@@ -1,5 +1,4 @@
 export * from './components';
-export * from './extend';
 export * from './foundation';
 export * from './learn';
 export * from './templates';
