+++ aries-site/src/examples/cardPreviews/data-how-to.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { Box, Button, TextInput } from 'grommet';
+import { Filter, Search } from 'grommet-icons';
+
+export const DataHowTo = () => (
+  <Box direction="row" align="center" gap="small">
+    <TextInput icon={<Search />} placeholder="Search" tabIndex={-1} />
+    <Button icon={<Filter />} kind="toolbar" tabIndex={-1} />
+  </Box>
+);
