+++ aries-site/src/examples/cardPreviews/index.js
@@ -6,6 +6,7 @@ export * from './checkbox';
 export * from './checkboxgroup';
 export * from './code-blocks';
 export * from './content-layouts';
+export * from './data-how-to';
 export * from './dateinput';
 export * from './footer';
 export * from './grid';
